refactored components files
